requirements.txt
requests
python-dotenv
win10toast